фильтр перенесен в метод
