+++ src/utils.js
@@ -25,32 +25,7 @@ const createElement = (template) => {
   return newElement.firstChild;
 };
 
-/* Функция отрисовки контента */
-const render = (root, content) => {
-  root.innerHTML = content;
-};
-
 /* Временная функция для проверки фильтров. Устанавливает время на 6 утра */
 const getTimeIsNow = () => moment().startOf(`day`).add(6, `hours`).format(`x`);
 
-const filterPoint = (data, filterName) => {
-  let sortData = [];
-  switch (filterName) {
-
-    case `filter-everything`:
-      sortData = data.events;
-      break;
-
-    case `filter-future`:
-
-      sortData = data.events.filter((it) => it.time[0] > getTimeIsNow());
-      break;
-
-    case `filter-past`:
-      sortData = data.events.filter((it) => it.time[0] < getTimeIsNow());
-  }
-  return sortData;
-};
-
-
-export {getRandomFromInterval, getRandomArr, getRandomDescription, createElement, getTimeIsNow, filterPoint, render};
+export {getRandomFromInterval, getRandomArr, getRandomDescription, createElement, getTimeIsNow};
